+++ src/main/resources/db/migration/V1__adding_tables.sql
@@ -3,7 +3,7 @@
 -- ============================================
 
 CREATE TABLE person (
-    id_person INT AUTO_INCREMENT PRIMARY KEY,
+    id_person BIGINT AUTO_INCREMENT PRIMARY KEY,
     email VARCHAR(255) NOT NULL UNIQUE,
     full_name VARCHAR(255) NOT NULL,
     phone_number VARCHAR(20) NOT NULL,
@@ -13,7 +13,7 @@ CREATE TABLE person (
 
 CREATE TABLE user (
     id_user BIGINT AUTO_INCREMENT PRIMARY KEY,
-    id_person INT NOT NULL,
+    id_person BIGINT NOT NULL,
     username VARCHAR(100) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
     identification_number VARCHAR(50) NOT NULL UNIQUE,
